Split out codec-only helpers from the full data pipeline helpers (codec + extstore).
