+++ packages/common/src/internal-non-workflow/codec-helpers.ts
@@ -1,16 +1,15 @@
+/**
+ * Codec-only helpers that operate on {@link Payload}s via {@link PayloadCodec}.
+ *
+ * For helpers that orchestrate the full data pipeline (converter + codec + extstore etc),
+ * see `./data-pipeline-helpers`.
+ *
+ * @module
+ */
 import type { temporal } from '@temporalio/proto';
 import type { Payload } from '../interfaces';
-import {
-  arrayFromPayloads,
-  convertOptionalToPayload,
-  fromPayloadsAtIndex,
-  toPayloadsWithContext,
-} from '../converter/payload-converter';
-import { PayloadConverterError } from '../errors';
 import type { PayloadCodec } from '../converter/payload-codec';
 import type { ProtoFailure } from '../failure';
-import type { LoadedDataConverter } from '../converter/data-converter';
-import type { UserMetadata } from '../user-metadata';
 import type { SerializationContext } from '../converter/serialization-context';
 import type {
   DecodedPayload,
@@ -19,8 +18,6 @@ import type {
   EncodedProtoFailure,
   ReplaceNested,
 } from './codec-types';
-import { runExternalRetrieve, runExternalStore } from './external-storage-runner';
-import { storageTargetFromContext } from '../converter/extstore';
 
 /** A `ProtoFailure` whose nested {@link Payload} fields all have type `P`. */
 type FailureWithPayloads<P extends Payload> = ReplaceNested<ProtoFailure, Payload, P>;
@@ -73,7 +70,7 @@ export async function decodeOptional(
   return await decode(codecs, payloads, context);
 }
 
-async function encodeSingle(
+export async function encodeSingle(
   codecs: PayloadCodec[],
   payload: Payload,
   context?: SerializationContext
@@ -111,84 +108,6 @@ export async function decodeOptionalSingle(
   return await decodeSingle(codecs, payload, context);
 }
 
-/** Run {@link PayloadCodec.decode} and convert from a single Payload */
-export async function decodeOptionalSinglePayload<T>(
-  dataConverter: LoadedDataConverter,
-  payload?: Payload | null | undefined,
-  context?: SerializationContext
-): Promise<T | null | undefined> {
-  if (payload == null) return payload;
-  const { payloadConverter, payloadCodecs, externalStorage } = dataConverter;
-  const [retrievedPayload] = await runExternalRetrieve({ externalStorage, payloads: [payload] });
-  const decodedPayload = await decodeOptionalSingle(payloadCodecs, retrievedPayload, context);
-  if (decodedPayload == null) return decodedPayload;
-  return payloadConverter.fromPayload(decodedPayload, context);
-}
-
-/**
- * Run {@link PayloadConverter.toPayload} on value, and then encode it.
- */
-export async function encodeToPayload(
-  converter: LoadedDataConverter,
-  value: unknown,
-  context?: SerializationContext
-): Promise<Payload> {
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  const encodedPayload = await encodeSingle(payloadCodecs, payloadConverter.toPayload(value, context), context);
-  const [storedPayload] = await runExternalStore({ externalStorage, target: storageTargetFromContext(context), payloads: [encodedPayload] });
-  return storedPayload!;
-}
-
-/**
- * Decode `payloads` and then return {@link arrayFromPayloads}`.
- */
-export async function decodeArrayFromPayloads(
-  converter: LoadedDataConverter,
-  payloads?: Payload[] | null,
-  context?: SerializationContext
-): Promise<unknown[]> {
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  const retrievedPayloads = payloads != null ? await runExternalRetrieve({ externalStorage, payloads }) : payloads;
-  return arrayFromPayloads(payloadConverter, await decodeOptional(payloadCodecs, retrievedPayloads, context), context);
-}
-
-/**
- * Decode `payloads` and then return {@link fromPayloadsAtIndex}.
- */
-export async function decodeFromPayloadsAtIndex<T>(
-  converter: LoadedDataConverter,
-  index: number,
-  payloads?: Payload[] | null,
-  context?: SerializationContext
-): Promise<T> {
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  const retrievedPayloads = payloads != null ? await runExternalRetrieve({ externalStorage, payloads }) : payloads;
-  return await fromPayloadsAtIndex(
-    payloadConverter,
-    index,
-    await decodeOptional(payloadCodecs, retrievedPayloads, context),
-    context
-  );
-}
-
-/**
- * Run extstore retrieve and codec decode over every payload field of a
- * {@link ProtoFailure}, then convert the result to an {@link Error}.
- */
-export async function decodeOptionalFailureToOptionalError(
-  converter: LoadedDataConverter,
-  failure: ProtoFailure | undefined | null,
-  context?: SerializationContext
-): Promise<Error | undefined> {
-  if (!failure) return undefined;
-  const { failureConverter, payloadConverter, payloadCodecs, externalStorage } = converter;
-  const decoded = await transformFailurePayloads(failure, async (payloads) => {
-    const retrieved = await runExternalRetrieve({ externalStorage, payloads });
-    return decode(payloadCodecs, retrieved, context);
-  });
-  return failureConverter.failureToError(decoded, payloadConverter, context);
-}
-
 export async function decodeOptionalMap(
   codecs: PayloadCodec[],
   payloads: Record<string, Payload> | null | undefined,
@@ -198,56 +117,6 @@ export async function decodeOptionalMap(
   return Object.fromEntries(
     await Promise.all(Object.entries(payloads).map(async ([k, v]) => [k, (await decode(codecs, [v], context))[0]]))
   );
-}
-
-/**
- * Run {@link PayloadConverter.toPayload} on values, and then encode them.
- */
-export async function encodeToPayloads(
-  converter: LoadedDataConverter,
-  ...values: unknown[]
-): Promise<Payload[] | undefined> {
-  return encodeToPayloadsWithContext(converter, undefined, values);
-}
-
-/**
- * Run {@link PayloadConverter.toPayload} on values with an optional serialization context, and then encode them.
- */
-export async function encodeToPayloadsWithContext(
-  converter: LoadedDataConverter,
-  context: SerializationContext | undefined,
-  values: unknown[]
-): Promise<Payload[] | undefined> {
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  if (values.length === 0) {
-    return undefined;
-  }
-  const payloads = toPayloadsWithContext(payloadConverter, context, values);
-  if (!payloads) return undefined;
-  const encoded = await encode(payloadCodecs, payloads, context);
-  return await runExternalStore({ externalStorage, target: storageTargetFromContext(context), payloads: encoded });
-}
-
-/**
- * Run {@link PayloadCodec.decode} and then {@link PayloadConverter.fromPayload} on values in `map`.
- */
-export async function decodeMapFromPayloads<K extends string>(
-  converter: LoadedDataConverter,
-  map: Record<K, Payload> | null | undefined,
-  context?: SerializationContext
-): Promise<Record<K, unknown> | undefined> {
-  if (!map) return undefined;
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  return Object.fromEntries(
-    await Promise.all(
-      Object.entries(map).map(async ([k, payload]): Promise<[K, unknown]> => {
-        const [retrieved] = await runExternalRetrieve({ externalStorage, payloads: [payload as Payload] });
-        const [decodedPayload] = await decode(payloadCodecs, [retrieved!], context);
-        const value = payloadConverter.fromPayload(decodedPayload!, context);
-        return [k as K, value];
-      })
-    )
-  ) as Record<K, unknown>;
 }
 
 /** Run {@link PayloadCodec.encode} on all values in `map` */
@@ -265,45 +134,6 @@ export async function encodeMap<K extends string>(
       })
     )
   ) as Record<K, EncodedPayload>;
-}
-
-/**
- * Run {@link PayloadConverter.toPayload} and then {@link PayloadCodec.encode} on values in `map`.
- */
-export async function encodeMapToPayloads<K extends string>(
-  converter: LoadedDataConverter,
-  map: Record<K, unknown>,
-  context?: SerializationContext
-): Promise<Record<K, Payload>> {
-  const { payloadConverter, payloadCodecs, externalStorage } = converter;
-  return Object.fromEntries(
-    await Promise.all(
-      Object.entries(map).map(async ([k, v]): Promise<[K, Payload]> => {
-        const payload = payloadConverter.toPayload(v, context);
-        if (payload === undefined) throw new PayloadConverterError(`Failed to encode entry: ${k}: ${v}`);
-        const [encodedPayload] = await encode(payloadCodecs, [payload], context);
-        const [storedPayload] = await runExternalStore({ externalStorage, target: storageTargetFromContext(context), payloads: [encodedPayload!] });
-        return [k as K, storedPayload!];
-      })
-    )
-  ) as Record<K, Payload>;
-}
-
-/**
- * Convert `error` to a {@link ProtoFailure}, then run codec encode + extstore
- * store over every payload.
- */
-export async function encodeErrorToFailure(
-  dataConverter: LoadedDataConverter,
-  error: unknown,
-  context?: SerializationContext
-): Promise<ProtoFailure> {
-  const { failureConverter, payloadConverter, payloadCodecs, externalStorage } = dataConverter;
-  const raw = failureConverter.errorToFailure(error, payloadConverter, context);
-  return await transformFailurePayloads(raw, async (payloads) => {
-    const encoded = await encode(payloadCodecs, payloads, context);
-    return runExternalStore({ externalStorage, target: storageTargetFromContext(context), payloads: encoded });
-  });
 }
 
 /**
@@ -385,54 +215,12 @@ export function noopDecodeMap<K extends string>(
   return map as Record<K, DecodedPayload> | null | undefined;
 }
 
-export async function encodeUserMetadata(
-  dataConverter: LoadedDataConverter,
-  staticSummary: string | undefined,
-  staticDetails: string | undefined,
-  context?: SerializationContext
-): Promise<temporal.api.sdk.v1.IUserMetadata | undefined> {
-  if (staticSummary == null && staticDetails == null) return undefined;
-
-  const { payloadConverter, payloadCodecs, externalStorage } = dataConverter;
-  const encodedSummary = await encodeOptionalSingle(
-    payloadCodecs,
-    convertOptionalToPayload(payloadConverter, staticSummary, context),
-    context
-  );
-  const encodedDetails = await encodeOptionalSingle(
-    payloadCodecs,
-    convertOptionalToPayload(payloadConverter, staticDetails, context),
-    context
-  );
-
-  if (encodedSummary == null && encodedDetails == null) return undefined;
-
-  const store = async (p: Payload | null | undefined) =>
-    p != null ? (await runExternalStore({ externalStorage, target: storageTargetFromContext(context), payloads: [p] }))[0] : undefined;
-  const [summary, details] = await Promise.all([store(encodedSummary), store(encodedDetails)]);
-  return { summary, details };
-}
-
-export async function decodeUserMetadata(
-  dataConverter: LoadedDataConverter,
-  metadata: temporal.api.sdk.v1.IUserMetadata | undefined | null,
-  context?: SerializationContext
-): Promise<UserMetadata> {
-  const res = { staticSummary: undefined, staticDetails: undefined };
-  if (metadata == null) return res;
-
-  const staticSummary =
-    (await decodeOptionalSinglePayload<string>(dataConverter, metadata.summary, context)) ?? undefined;
-  const staticDetails =
-    (await decodeOptionalSinglePayload<string>(dataConverter, metadata.details, context)) ?? undefined;
-
-  return { staticSummary, staticDetails };
-}
-
 /**
  * Walk every payload field of a {@link ProtoFailure} and apply a `transform`.
+ *
+ * @internal
  */
-async function transformFailurePayloads<P extends Payload>(
+export async function transformFailurePayloads<P extends Payload>(
   failure: ProtoFailure,
   transform: (payloads: Payload[]) => Promise<P[]>
 ): Promise<FailureWithPayloads<P>> {
@@ -478,6 +266,6 @@ async function transformFailurePayloads<P extends Payload>(
         }
       : undefined,
   };
-  
+
   return result as FailureWithPayloads<P>;
 }
